requirements.txt
pymongo==3.6.0
confluent-kafka==0.11.0
PyYAML==3.12
